Refactor send_msg function to include authentication
The send_msg function in the client module has been modified to include authentication. Along with this change, corresponding adjustments were made to related functions and method calls in "client.rs" and "main.rs". Now, the 'auth' argument is retrieved from the get_vars function and passed to send_msg.